Segments abstracting
